Update TechOrder Manager to allow for multiple
